requirements.txt
# requirements.txt
numpy==1.21.6
opencv-python==4.7.0
imutils==0.5.4
scikit-image==0.19.3
tensorflow==2.9.1
py-sudoku==1.0.3
Flask==2.0.1
gunicorn==20.1.0
python-dotenv==0.19.0